refactor: simplify
