+++ apps/desktop/src/entities/user/components/user-avatar.tsx
@@ -1,18 +1,12 @@
-import { UserAvatar as UIUserAvatar } from '@conar/ui/components/user-avatar'
+import { UserAvatar as UIUserAvatar } from '@conar/ui/components/custom/user-avatar'
 import { authClient } from '~/lib/auth'
 
-interface UserAvatarProps {
-  className?: string
-  fallbackClassName?: string
-}
-
-export function UserAvatar({ className, fallbackClassName }: UserAvatarProps) {
+export function UserAvatar({ className }: { className?: string }) {
   const { data } = authClient.useSession()
   return (
     <UIUserAvatar
       user={data?.user}
       className={className}
-      fallbackClassName={fallbackClassName}
     />
   )
 }
