+++ apps/web/src/routes/account.tsx
@@ -3,11 +3,11 @@ import { AppLogo } from '@conar/ui/components/brand/app-logo'
 import { Button } from '@conar/ui/components/button'
 import { LoadingContent } from '@conar/ui/components/custom/loading-content'
 import { ThemeToggle } from '@conar/ui/components/custom/theme-toggle'
+import { UserAvatar } from '@conar/ui/components/custom/user-avatar'
 import { Dialog, DialogClose, DialogContent, DialogFooter, DialogHeader, DialogTitle, DialogTrigger } from '@conar/ui/components/dialog'
 import { Label } from '@conar/ui/components/label'
 import { Separator } from '@conar/ui/components/separator'
 import { Textarea } from '@conar/ui/components/textarea'
-import { UserAvatar } from '@conar/ui/components/user-avatar'
 import { cn } from '@conar/ui/lib/utils'
 import NumberFlow from '@number-flow/react'
 import { RiDashboard3Line, RiFileListLine, RiGitBranchLine, RiGithubFill, RiLogoutCircleLine, RiMessageLine, RiMoonLine, RiSunLine } from '@remixicon/react'
@@ -242,7 +242,7 @@ function AccountLayout() {
         <aside className="w-64 shrink-0">
           <div className="mb-4 flex items-center gap-3">
             <UserAvatar user={user} className="size-10 shrink-0" />
-            <div className="min-w-0 flex-1 flex flex-col">
+            <div className="flex min-w-0 flex-1 flex-col">
               <p className="truncate font-medium">{user.name}</p>
               <p className="truncate text-sm text-muted-foreground">{user.email}</p>
             </div>
